change the match id list to a set
we need a set because the match history api will return a list of 25 match_ids that start with the last match id that you have, meaning that the first entry in those 25 match ids will match the last entry of your previous 25 match_ids.  in python, sets are iterable and have a length property like lists, so it's safe to make the change.